pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage"
version = "1.0.0"
description = "SupportTriage OpenEnv environment"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "openenv-support-triage" }
]
dependencies = [
  "openenv>=0.2.0",
  "fastapi==0.111.0",
  "uvicorn[standard]==0.29.0",
  "pydantic==2.7.1",
  "pyyaml==6.0.1",
  "openai>=1.30.0",
  "requests>=2.31.0",
  "python-multipart==0.0.9"
]

[project.scripts]
server = "app:run_server"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "tasks*", "graders*", "baseline*"]